🚸 implement Logout
Added a logout() method to the frontend code, which sends a POST request to the logout endpoint (this.apiUrlLogout). The request includes null as the request body and { withCredentials: true } to ensure the credentials (cookies) are sent. Upon a successful response, the isLoggedIn() method is invoked to update the user's login status.